added vite bundler
